+++ vnstock/connector/dnse/__init__.py
@@ -1,1 +1,6 @@
-from .trade import Trade  # noqa: F401
+"""
+DNSE broker connector removed.
+
+The DNSE broker execution connector (login, orders, account management)
+was removed as part of the data-only package refactor.
+"""