+++ vnstock/ui/domains/broker/__init__.py
@@ -1,5 +1,5 @@
 """
-DNSE broker domain removed.
+Broker domain removed.
 
 DNSE account/order execution was removed as part of the data-only package refactor.
 """